[refactor] Move LockTable.LockType to Lock.LockType
