+++ src/org/exist/storage/lock/MultiReadReentrantLock.java
@@ -39,7 +39,7 @@ package org.exist.storage.lock;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.exist.storage.lock.LockTable.LockType;
+import org.exist.storage.lock.Lock.LockType;
 import org.exist.util.DeadlockException;
 import org.exist.util.LockException;
